Blueprinted a Mute / Unmute Button & Blueprint.
For the time being, they are set to Thumbnail left on the VR hand controller or left mouse click on a mouse.

The music is just a random clip that I have downloaded from the internet, however we can search for a better song to suit our OFFICE / HAMMER theme.